refactor: Move image and marker constants to mapConstants for better organization
