+++ src/platform/src/common/features/airQuality-map/constants/mapConstants.js
@@ -8,6 +8,17 @@ import LightMode from '@/images/map/light.webp';
 import SatelliteMode from '@/images/map/satellite.webp';
 import StreetsMode from '@/images/map/street.webp';
 
+import { renderToString } from 'react-dom/server';
+
+// Import icon components
+import GoodAir from '@/icons/Charts/GoodAir';
+import ModerateAir from '@/icons/Charts/Moderate';
+import UnhealthyForSensitiveGroups from '@/icons/Charts/UnhealthySG';
+import Unhealthy from '@/icons/Charts/Unhealthy';
+import VeryUnhealthy from '@/icons/Charts/VeryUnhealthy';
+import Hazardous from '@/icons/Charts/Hazardous';
+import Invalid from '@/icons/Charts/Invalid';
+
 export const BOUNDARY_URL = 'https://nominatim.openstreetmap.org/search';
 
 export const mapStyles = [
@@ -87,4 +98,89 @@ export const LAYOUT_CONFIG = {
     SIDEBAR_MIN_WIDTH: '380px',
     SIDEBAR_LG_WIDTH: '470px',
   },
+};
+
+// -------------------------------------------------------------------
+// Icon Images: Encoded SVG images for use in markers/popups
+// -------------------------------------------------------------------
+export const images = {
+  GoodAir: `data:image/svg+xml,${encodeURIComponent(renderToString(<GoodAir />))}`,
+  ModerateAir: `data:image/svg+xml,${encodeURIComponent(renderToString(<ModerateAir />))}`,
+  UnhealthyForSensitiveGroups: `data:image/svg+xml,${encodeURIComponent(renderToString(<UnhealthyForSensitiveGroups />))}`,
+  Unhealthy: `data:image/svg+xml,${encodeURIComponent(renderToString(<Unhealthy />))}`,
+  VeryUnhealthy: `data:image/svg+xml,${encodeURIComponent(renderToString(<VeryUnhealthy />))}`,
+  Hazardous: `data:image/svg+xml,${encodeURIComponent(renderToString(<Hazardous />))}`,
+  Invalid: `data:image/svg+xml,${encodeURIComponent(renderToString(<Invalid />))}`,
+  undefined: `data:image/svg+xml,${encodeURIComponent(renderToString(<Invalid />))}`,
+};
+
+// -------------------------------------------------------------------
+// Marker thresholds and colors based on pollutant type
+// -------------------------------------------------------------------
+export const markerDetails = {
+  pm2_5: [
+    { limit: 500.5, category: 'Invalid' },
+    { limit: 225.5, category: 'Hazardous' },
+    { limit: 125.5, category: 'VeryUnhealthy' },
+    { limit: 55.5, category: 'Unhealthy' },
+    { limit: 35.5, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 9.1, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+  pm10: [
+    { limit: 604.1, category: 'Invalid' },
+    { limit: 424.1, category: 'Hazardous' },
+    { limit: 354.1, category: 'VeryUnhealthy' },
+    { limit: 254.1, category: 'Unhealthy' },
+    { limit: 154.1, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 54.1, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+  no2: [
+    { limit: 2049.1, category: 'Invalid' },
+    { limit: 1249.1, category: 'Hazardous' },
+    { limit: 649.1, category: 'VeryUnhealthy' },
+    { limit: 360.1, category: 'Unhealthy' },
+    { limit: 100.1, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 53.1, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+  o3: [
+    { limit: 604.1, category: 'Invalid' },
+    { limit: 504.1, category: 'Hazardous' },
+    { limit: 404.1, category: 'VeryUnhealthy' },
+    { limit: 204.1, category: 'Unhealthy' },
+    { limit: 154.1, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 54.1, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+  co: [
+    { limit: 50.5, category: 'Invalid' },
+    { limit: 40.5, category: 'Hazardous' },
+    { limit: 30.5, category: 'VeryUnhealthy' },
+    { limit: 10.5, category: 'Unhealthy' },
+    { limit: 4.5, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 2.5, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+  so2: [
+    { limit: 1004.1, category: 'Invalid' },
+    { limit: 804.1, category: 'Hazardous' },
+    { limit: 604.1, category: 'VeryUnhealthy' },
+    { limit: 304.1, category: 'Unhealthy' },
+    { limit: 185.1, category: 'UnhealthyForSensitiveGroups' },
+    { limit: 75.1, category: 'ModerateAir' },
+    { limit: 0.0, category: 'GoodAir' },
+  ],
+};
+
+export const colors = {
+  Invalid: '#C6D1DB',
+  Hazardous: '#D95BA3',
+  VeryUnhealthy: '#AC5CD9',
+  Unhealthy: '#F7453C',
+  UnhealthyForSensitiveGroups: '#FF851F',
+  ModerateAir: '#FFD633',
+  GoodAir: '#34C759',
+  undefined: '#C6D1DB',
 };
